Added GameBoardTest
Currently, this method only tests the getClosest and getRandom methods.

In the future, I am going to move the getClosest/Random methods out of
Bug, and replace them with these. That way they are in only one place
and can be optimized.

Additionally, I am going to make the Map a Singleton -- that way, we can
stop passing it around as a parameter everywhere.